requirements.txt
scikit-learn>=1.0.2
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
joblib>=1.3.0
seaborn>=0.12.0
speedtest-cli>=2.1.3
ping3>=4.0.0
psutil>=5.8.0
requests>=2.26.0
python-dotenv>=0.19.0
geocoder>=1.38.1 